feat: filter options
